+++ otter/src/system.rs
@@ -24,7 +24,7 @@ pub enum ThreadType {
 /// Wraps `enif_thread_type`.
 pub fn thread_type() -> ThreadType {
     use crate::sys::{NIF_THR_UNDEFINED, NIF_THR_NORMAL_SCHEDULER, NIF_THR_DIRTY_CPU_SCHEDULER, NIF_THR_DIRTY_IO_SCHEDULER};
-    match crate::wrapper::system::thread_type() {
+    match unsafe { crate::enif::thread_type() } {
         NIF_THR_UNDEFINED          => ThreadType::NonScheduler,
         NIF_THR_NORMAL_SCHEDULER   => ThreadType::Scheduler,
         NIF_THR_DIRTY_CPU_SCHEDULER => ThreadType::DirtyCpu,
@@ -37,5 +37,5 @@ pub fn thread_type() -> ThreadType {
 ///
 /// Wraps `enif_system_info`.
 pub fn system_info(info: &mut SysInfo) {
-    crate::wrapper::system::system_info(info);
+    unsafe { crate::enif::system_info(info, std::mem::size_of::<SysInfo>()) };
 }